requirements.txt
playwright==1.58.0
pytest==8.3.5
pytest-html==4.1.1
allure-pytest==2.13.5
pytest-xdist==3.6.1
pytest-ordering==0.6
pyyaml==6.0.2
faker==26.0.0
